separate parent bt class
